Move StorageAuthorityCheckAlgorithm
