+++ shardingsphere-features/shardingsphere-authority/shardingsphere-authority-common/src/main/resources/META-INF/services/org.apache.shardingsphere.authority.spi.AuthorityCheckAlgorithm
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.authority.engine.impl.StorageAuthorityCheckAlgorithm
+org.apache.shardingsphere.authority.algorithm.StorageAuthorityCheckAlgorithm
